+++ src/main/java/com/after_sunrise/cryptocurrency/bitflyer4j/entity/OrderList.java
@@ -71,7 +71,7 @@ public interface OrderList {
         private final String parentId;
 
         @SerializedName("count")
-        private final Long count;
+        private final Integer count;
 
         @SerializedName("before")
         private final Long before;
